requirements.txt
fastapi
strawberry-graphql
strawberry-graphql[debug-server]
uvicorn
sqlalchemy
asyncio
